+++ 06_gpu_and_ml/llm-serving/sglang_kitchen_sink.py
@@ -1,6 +1,6 @@
 # ---
 # deploy: true
-# cmd: ["python", "06_gpu_and_ml/llm-serving/sglang_kitchen_sink.py"]
+# cmd: ["python", "06_gpu_and_ml/llm-serving/sglang_kitchen_sink_server.py"]
 # ---
 
 # # Fast-booting, low-latency Qwen 3 8B with SGLang, GPU snapshots, and speculative decoding
@@ -21,7 +21,6 @@ import time
 
 import aiohttp
 import modal
-import modal.experimental
 
 MINUTES = 60  # seconds
 
@@ -31,7 +30,7 @@ sglang_image = (
     ).entrypoint([])  # silence chatty logs on container start
 )
 
-sglang_image.env(
+sglang_image = sglang_image.env(
     {  # bleeding-edge SGLang perf opt settings
         "SGLANG_ENABLE_SPEC_V2": "1",
         "SGLANG_ENABLE_DFLASH_SPEC_V2": "1",
@@ -132,11 +131,11 @@ def wake_up():
 # ### Selecting infrastructure to minimize latency
 
 REGION = "us"
-PROXY_REGION = "us-west"
+ROUTING_REGION = "us-west"
 
 MIN_CONTAINERS = 0  # set to 1 to ensure one replica is always ready
 
-# ### Determining autoscaling policy with `@modal.concurrent`
+# ### Determining autoscaling policy with `target_concurrency`
 
 TARGET_INPUTS = 10
 
@@ -168,7 +167,7 @@ app = modal.App(name="example-sglang-kitchen-sink")
 PORT = 8000
 
 
-@app.cls(
+@app._experimental_server(
     image=sglang_image,
     gpu=GPU,
     volumes={DG_CACHE_PATH: DG_CACHE_VOL, HF_CACHE_PATH: HF_CACHE_VOL},
@@ -176,13 +175,11 @@ PORT = 8000
     min_containers=MIN_CONTAINERS,
     enable_memory_snapshot=True,
     experimental_options={"enable_gpu_snapshot": True},
-)
-@modal.experimental.http_server(
     port=PORT,  # wrapped code must listen on this port
-    proxy_regions=[PROXY_REGION],  # location of proxies, should be same as Cls region
+    routing_regions=[ROUTING_REGION],  # location of proxies, should be same as Cls region
     exit_grace_period=15,  # seconds, time to finish up requests when closing down
+    target_concurrency=TARGET_INPUTS,
 )
-@modal.concurrent(target_inputs=TARGET_INPUTS)
 class SGLang:
     @modal.enter(snap=True)
     def startup(self):
@@ -240,19 +237,19 @@ class SGLang:
 # ## Deploy the server
 
 # ```bash
-# modal deploy sglang_kitchen_sink.py
+# modal deploy sglang_kitchen_sink_server.py
 # ```
 
 # ## Test the server
 
 # ```bash
-# modal run sglang_kitchen_sink.py
+# modal run sglang_kitchen_sink_server.py
 # ```
 
 
 @app.local_entrypoint()
 async def test(test_timeout=10 * MINUTES, prompt=None, twice=True):
-    url = SGLang._experimental_get_flash_urls()[0]
+    url = (await SGLang.get_urls.aio())[ROUTING_REGION]
 
     system_prompt = {
         "role": "system",
@@ -347,16 +344,16 @@ async def _send_request_streaming(
 # ### Test memory snapshotting
 
 # ```bash
-# python sglang_kitchen_sink.py
+# python sglang_kitchen_sink_server.py
 # ```
 
 if __name__ == "__main__":
     # after deployment, we can use the class from anywhere
-    SGLang = modal.Cls.from_name("example-sglang-kitchen-sink", "SGLang")
+    sglang_server = modal.Server.from_name("example-sglang-kitchen-sink", "SGLang")
 
     print("calling inference server")
     try:
-        asyncio.run(probe(SGLang._experimental_get_flash_urls()[0]))
+        asyncio.run(probe(sglang_server.get_urls()[ROUTING_REGION]))
     except modal.exception.NotFoundError as e:
         raise Exception(
             f"To take advantage of GPU snapshots, deploy first with modal deploy {__file__}"
